+++ app/src/main/AndroidManifest.xml
@@ -37,6 +37,10 @@
             android:name=".activities.BookmarkUsersActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar" />
+        <activity
+            android:name=".activities.SettingActivity"
+            android:label="@string/app_name"
+            android:theme="@style/AppTheme.NoActionBar" />
     </application>
 
 </manifest>
