+++ app/src/main/java/me/rei_m/hbfavmaterial/activities/BaseActivityWithDrawer.kt
@@ -1,0 +1,70 @@
+package me.rei_m.hbfavmaterial.activities
+
+import android.os.Bundle
+import android.support.design.widget.NavigationView
+import android.support.v4.view.GravityCompat
+import android.support.v4.widget.DrawerLayout
+import android.support.v7.app.ActionBarDrawerToggle
+import android.support.v7.app.AppCompatActivity
+import android.support.v7.widget.Toolbar
+import android.view.MenuItem
+import me.rei_m.hbfavmaterial.R
+import me.rei_m.hbfavmaterial.events.EventBusHolder
+
+public abstract class BaseActivityWithDrawer : AppCompatActivity(),
+        NavigationView.OnNavigationItemSelectedListener {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+        val toolbar = findViewById(R.id.toolbar) as Toolbar
+        setSupportActionBar(toolbar)
+
+        val drawer = findViewById(R.id.drawer_layout) as DrawerLayout
+
+        val toggle = ActionBarDrawerToggle(this,
+                drawer,
+                toolbar,
+                R.string.navigation_drawer_open,
+                R.string.navigation_drawer_close)
+
+        drawer.setDrawerListener(toggle)
+        toggle.syncState()
+
+        val navigationView = findViewById(R.id.nav_view) as NavigationView
+        navigationView.setNavigationItemSelectedListener(this)
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        // EventBus登録
+        EventBusHolder.EVENT_BUS.register(this)
+    }
+
+    override fun onPause() {
+        super.onPause()
+
+        // EventBus登録解除
+        EventBusHolder.EVENT_BUS.unregister(this)
+    }
+
+    override fun onBackPressed() {
+        val drawer = findViewById(R.id.drawer_layout) as DrawerLayout
+        if (drawer.isDrawerOpen(GravityCompat.START)) {
+            drawer.closeDrawer(GravityCompat.START)
+        } else {
+            super.onBackPressed()
+        }
+    }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    override fun onNavigationItemSelected(item: MenuItem): Boolean {
+
+        val drawer = findViewById(R.id.drawer_layout) as DrawerLayout
+
+        drawer.closeDrawer(GravityCompat.START)
+
+        return true
+    }
+}
