+++ app/src/main/java/me/rei_m/hbfavmaterial/activities/SettingActivity.kt
@@ -1,0 +1,33 @@
+package me.rei_m.hbfavmaterial.activities
+
+import android.content.Context
+import android.content.Intent
+import android.os.Bundle
+import android.support.v4.view.GravityCompat
+import android.support.v4.widget.DrawerLayout
+import android.view.MenuItem
+import me.rei_m.hbfavmaterial.R
+
+public class SettingActivity : BaseActivityWithDrawer() {
+
+    companion object {
+        public fun createIntent(context: Context): Intent {
+            return Intent(context, SettingActivity::class.java)
+        }
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        // TODO Fragmentを作成
+    }
+
+    @SuppressWarnings("StatementWithEmptyBody")
+    override fun onNavigationItemSelected(item: MenuItem): Boolean {
+
+        val drawer = findViewById(R.id.drawer_layout) as DrawerLayout
+
+        drawer.closeDrawer(GravityCompat.START)
+
+        return true
+    }
+}
