+++ app/src/main/res/layout/app_bar_main.xml
@@ -27,4 +27,12 @@
         android:layout_height="match_parent"
         android:paddingTop="?attr/actionBarSize" />
 
+    <FrameLayout
+        android:id="@+id/content"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:paddingTop="?attr/actionBarSize"
+        android:visibility="gone"
+        tools:ignore="MergeRootFrame" />
+
 </android.support.design.widget.CoordinatorLayout>
